+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 * `UploaderComponent::error()` renamed as `getError()`;
 * `Shell::folderChmod()` has a default value for the `$chmod` argument;
 * added `OptionsParser::consume()` method;
-* `TestCaseTrait::assertLogContains()` can now take absolute paths and filename
-    with extension;
+* added `TestCaseTrait::getLogFullPath()` method. So now `assertLogContains()`
+    and `deleteLog()` methods can take absolute paths and filenames with or
+    without extension;
 * fixed code and output for the `Shell` class;
 * removed `TestCaseTrait::deleteAllLogs()` method;
 * removed `folderIsWriteable()` global function. Use instead